refactor : header , background
